QcomModulePkg: BootLib: skip ValidateSlotGuids()
When using community builds, we want to depend as less as possible on
Qualcomm internal builds quirks. This function does a few checks which
are relevant when booting a Qualcomm Android image, but not well
suited for community builds. For example users might choose to remove
system_x partition, in which case the bootloader wouldn't boot
anymore.

Signed-off-by: Nicolas Dechesne <nicolas.dechesne@linaro.org>