QcomModulePkg: Fastboot: Add extend Fastboot menu to UART interface
We can already somewhat control the fastboot menu
over the serial UART, as up and down changes action mode.

However, there's no way to select the mode over the UART.

So add some feedback over the UART as to the current mode
and use SCAN_RIGHT as the selection option (equivalent to
the power button).

Signed-off-by: John Stultz <john.stultz@linaro.org>

[AmitP: It is equivalent of using VOL+ (up), VOL- (down)
        and POWER keys on db845c to scroll thru or select
        fastboot options displayed over HDMI display.]
Tested-by: Amit Pundir <amit.pundir@linaro.org>
Signed-off-by: Nicolas Dechesne <nicolas.dechesne@linaro.org>