QcomModulePkg: BootLib: set DTB best match explicitly
GetSocDtb() match and load the best DTB in dtb.img.
This match is done based on non-upstream DT properties
i.e. qcom,{msm-id/board-id/pmic-id}, which are not
supported on db845c.

Since we do not support multiple DTBs in dtb.img anyway,
we set the current (and only available dtb in dtb.img)
as the best match explicitly.

Signed-off-by: Amit Pundir <amit.pundir@linaro.org>