libavb: load and verify of boot image using multithreading
The load and verification of the boot image is split into
two segments. The load operation will be run by Load Thread on
Core 0 and Verify operation will be run by Verify Thread on
Core 7.

Semaphores and Mutex are used for synchronization. While
Load Thread is loading first half of the boot image
verify thread will wait on Semaphore Semloadfirst. Once the
Load Operation of first half of the image is completed load
thread signals semaphore Semloadfirst and verify thread will
continue to verify the first half of the boot image and load
thread will be loading second half of boot image parallely.
Once the second image is loaded the load thread will signal
the semaphore SemLoadSecond and the verify thread will
proceed to verify the second half of the boot image.

Change-Id: If370e2b0175556ffde8c3d43a6dbe33f1c31c881
Signed-off-by: Aravind Vijayakumar <aprasann@codeaurora.org>