QcomModulePkg: add condition to strictly patch last partition
Currently patching for the last and last to second partition is causing incorrect size
for last to second partition.

As a fix:
Adding a condition to differentiate between last patition of a LUN with GUID Type="0"
and PartionName="last_parti" vs an auto grow last partition with any other name except
"last_parti". This will help to know last partition of any LUN and patch accordingly.

Change-Id: I9ee7523d6f0209d9da00340b7989927dfec6e7e7