libavb: load and verify of boot image using multithreading
The load and verification of the boot image is split into
two segments. The load operation will be run by Load Thread on
Core 0 and Verify operation will be run by Verify Thread on
Core 7.

Semaphores are used for synchronization. While the
load thread is loading first half of the boot image
verify thread will wait on semaphore Semloadfirst. Once the
load operation of first half of the image is completed load
thread signals semaphore Semloadfirst and verify thread will
continue to verify the first half of the boot image and load
thread will be loading second half of boot image parallely.
Once the second image is loaded the load thread will signal
the semaphore SemLoadSecond and the verify thread will
proceed to verify the second half of the boot image.

Change-Id: Id6292c0a4a62422b2427a09aeb95f58d418fe880
Signed-off-by: Aravind Vijayakumar <quic_aprasann@quicinc.com>