QcomModulePkg: set SingleDtHdr from ImageBuffer
Set the pointer SingleDtHdr at the relevant position relative to DtbOffset.

Depending on the boot-header version, ImageBuffer is computed differently with
DtbOffset. E.g, v2 will set ImageBuffer to the beginning of the boot-image and
account for the boot-header size; v3 will set ImageBuffer in the vendor-image
and DtbOffset will account for its header; v0/v1 will set ImageBuffer at the
start of the kernel image, hence after the boot-header and DtbOffset
(calculated originally in GZipPkgCheck) will remain relative to the kernel it
was appended to, itself part of the whole boot-image.

Change-Id: I5cc2ef178446298af313cabffeab3e7c5dfe139a
Signed-off-by: Parikshit Pareek <quic_ppareek@quicinc.com>