edk2: Add support for SCMI channel and DT node updates for SCMI instances
This change adds support for populating SCMI shared memory channel
and patching of HLOS DT. This is in preparation of SCMI migration.
This change looks for scmi dt nodes with specific compatible string
and updates the IRQ information in that node which it fetches from RM
via UEFI RmVm driver API.

Additionally, it also populate the SCMI shmem channel with capability-id
of the outgoing doorbell and the constant function-id. These values are
in turn used by the HLOS SCMI transport driver.

Change-Id: I5fec20d0519d49696b4968560dfe23182a24b360
Signed-off-by: Nikunj Kela <quic_nkela@quicinc.com>