edk2: ABL shall not add "ro" if boot image contains "rw"
Previously ABL added "ro" to kernel cmdline by default. Now if ABL
detects "rw" in the boot image cmdline, ABL will not add "ro".

This allows the flexibility at build-time to have root mounted as
read-write, while maintaining the default of read-only.

This functionality is needed for ostree systems that require the root
device to be mounted as read-write.

Change-Id: I32661591f093f27a20fee3cd9981e58494cbcf5e
Signed-off-by: Eric Smith <quic_ericsmit@quicinc.com>