QcomModulePkg: edk2: read the cell counts from DT
Currently it is assumed that cells count for address and size for
SCMI shmem subnodes will always be 2 which is not always true.
Even on 64bit architecture, SCMI channels can use cell count as 1
if the shmem addresses are < 4GB. Therefore, use scmichannels alias
to get the cell count efficiently.

Change-Id: Ic22874a472ef08b17e758a7fa4ed06d1c04459a7