QcomModulePkg: Update number of partition on target
The maximum number of partitions that is supported in a
LUN is 128. As the number of partitions enumerated on
targets are more than 128, the calculation needs to be
updated. The maximum number of partitions that is
supported in the system is calculated by

 (MAX_NUM_PARTITIONS_PER_LUN * MAX_LUNS)

where MAX_NUM_PARTITIONS_PER_LUN is 128.

Change-Id: Ie1880893de823222d2c84dad2efc3add32ae9a09