QcomModulePkg: Disable dtbo overlay.
PVM is not using dtbo partition but if dtbo is flashed, bootloader will
try to overlay the device tree with the content of dtbo image and it
will lead to undesired behavior of the guest(most likely the result
will be crash of the guest). So, explicitly disable the dtbo need from
bootloader.

Change-Id: I85480bd8492940677b81009c251beaac61643edc
Signed-off-by: Ankurkumar Satasiya <asatasiy@qti.qualcomm.com>